minor changes based on the latest API specs

+++ src/main/java/io/contek/invoker/deribit/api/common/_Deposit.java
@@ -6,7 +6,7 @@ import javax.annotation.concurrent.NotThreadSafe;
 public class _Deposit {
 
   public String address;
-  public Long amount;
+  public Double amount;
   public String currency;
   public Long received_timestamp;
   public String state;
